server_FastAPI/requirements.txt
fastapi
uvicorn
motor
python-dotenv
pymongo
bcrypt
python-jose[cryptography]
cloudinary
python-multipart
pydantic-settings
jwt